move encodings to folder and test rett
