+++ python/ferrmion/encoding/ternary_tree.py
