+++ python/ferrmion/encoding/ternary_tree_node.py
